Replaced `hide()` methods with CSS workaround
We are not using display:none here because of the `required`
attributes. If you have an input with a `required` attribute
and you add chosen to it, using the `hide()` method won't
allow you to save your form and YOU WON'T be informed.

Using this workaround the input is not visible but it's still
there, allowing the browser to display the error.